Authorized admin route for pix master (and not the normal route)
